Refactor remove unecesary Contribute module
